requirements.txt
websockets
Faker
httpx
python-okx
